packages/agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langgraph==0.2.24
langchain==0.2.14
google-generativeai==0.7.2
pydantic==2.8.2
httpx==0.27.0
psycopg[binary]==3.2.1
